Update README.md
This research focuses on the implementation and optimization of Huffman encoding in order to improve the efficiency of information communication. The project's goal is to enhance channel utilization, reduce transmission time, and minimize costs by using Huffman trees and data structures. The study looks on efficient Huffman tree construction techniques for increased compression efficiency, as well as dynamic Huffman coding scenarios for flexibility to changing data frequencies. It also investigates error detection and correction strategies to improve Huffman coding/decoding resilience. The use of the Huffman algorithm in this methodology has produced encouraging results, including increased compression efficiency, robust adaptation to dynamic settings, and improved error resilience. This study makes a substantial contribution to the area by emphasizing the practical consequences and importance of data compression and encoding in real-world applications.